[roll] Roll fuchsia [developer] Always use in-development level for protocol compatibility check
Previously, Fuchsia builds used `__Fuchsia_API_level__`, which will
never match `CURRENT_SUPPORTED_API_LEVEL` in platform builds. To fix
this, use the same build variable as used for host tool builds.

Rename FUCHSIA_API_LEVEL to NEXT_STABLE_API_LEVEL to avoid confusion
and prepare for `NEXT` when the in-development level will not be
numbered. This name is consistent with the value of the GN variable
`platform_version.in_development_api_level` from which it gets its
value and avoids needing to adjust any logic in the header.

Also ensure NEXT_STABLE_API_LEVEL is not less than
`CURRENT_SUPPORTED_API_LEVEL` because the logic does not allow this.

This fixes the naming issue described in the bug. This entire mechanism
will be replaced in the future.

Original-Bug: 308808613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063701
Original-Revision: 1e5e9b4f94f0276fbd11a1695599106dc8d7b143
GitOrigin-RevId: 24d4e06b3a5edec3c65d9e9f86ade86dc0d706b1
Change-Id: I1cfc272495b732e381db7329cb0cb6b17249006a