[roll] Roll fuchsia [bazel] pass compilation mode to refresh-compdb
The refresh-bazel-compdb command was not taking the compilation mode
into account when running the query which meant that all of the paths
were pointing to the fastbuild directories which do not exist in most
builds. This CL figures out what compilation mode you are using for the
normal build and mimics it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064437
Original-Revision: 61b53f06d65819eaf1ceb2135267dfd2ea38cea8
GitOrigin-RevId: b816390b7811ce617d12b95954d5f4d3afa3398c
Change-Id: Ic5032dc0e01da24b9ab5b980a110f712c4a9bc14