[roll] Roll fuchsia [kernel] Refactor TimerQueue in preparation for boottime
This CL refactors TimerQueue::Tick() to call a new TickInternal()
method that, in addition to the current timestamp and CPU that Tick
takes, also takes a reference to the timer_list_ to operate on.

This is a no-op today, but will be useful in the future when the
TimerList contains two different timer lists (one for the boottime
timeline and one for monotonic).

Because the two timelines technically operate on different time types
(zx_time_t and zx_boot_time_t), TickInternal is templated on this type.
Of course, given that all of these time types are typedefs of the same
underlying type, this won't matter today. However, this will matter in
the future when stronger time types are in use. Moreover, it provides
clarity to callers that the type of the `now` argument is not always
zx_time_t.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060142
Original-Revision: 6dae1fb22af4363c0a68fffcd655670744d67bb3
GitOrigin-RevId: f50b3cea5094ddd35e064a6f156d730ca8dc586e
Change-Id: I236b173f9aa05928af1fcf050a98780246851618