[roll] Roll fuchsia [utest][core] Remove tests with debug dependencies
Removes a couple of tests that were using kernel shell commands for
functionality. These shell commands are not an API, subject to change,
and not always available depending on the build of the kernel.

The functionality these tests are attempting to validate is either
already tested as best as possible with in kernel tests, or is not
behavior that is guaranteed and predictable and therefore cannot have
a reliable test.

Original-Bug: 340232040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064072
Original-Revision: 41eb7319f951302599e051d0932a5f05956a608b
GitOrigin-RevId: 5b5a7648a3107a305aadd60c0e7acedede783ab4
Change-Id: Ib082bdff94b41ff1c6e973a1dfd58f0269317e21