[roll] Roll fuchsia [sdk][build] Stop processing //sdk/BUILD.gn with toolchain variants
A dependency on //sdk/BUILD.gn was added to the Go version-history
library in https://fxrev.dev/1043572. Because the library's build file
is processed with target (Fuchsia) toolchain variants, this resulted in
//sdk/BUILD.gn also being processed with such variants. This is
unnecessary and can create issues for SDK GN templates that do not
expect this. (For example, this new dependency caused
https://fxrev.dev/1058658 to start failing.)

The version history does not differ by toolchain, so change the
dependency to the `default_toolchain` to avoid processing //sdk/BUILD.gn
with other toolchains.

To prevent such issues in the future, only define //sdk:version_history
in the `default_toolchain` and add an assert //sdk/BUILD.gn to prevent
such dependencies from being unintentionally introduced in the future.

Original-Bug: 344633975, 324892812
Test: core.x64-asan bot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061072
Original-Revision: 9d8cf016421cb53344a20df6c260d2853300cd01
GitOrigin-RevId: 825ab251331f69ca6f3d939f5c3bd87d464ca996
Change-Id: Id8bf18eddaf51eab9516265439f18676d3e32e5e