[roll] Roll infra/recipes [testing_requests] Fix Integration detection for try builds
testing_requests was using the gitiles commit to detect the current
commit host and then using that determine which integration repo it
was dealing with. However, on a try build, this doesn't work because the
gitiles commit points to the repo of the commit being tried, e.g.
the fuchsia repo, not the integration repo.

Instead, we propagate the full build spec (instead of just spec.test)
and check spec.test.remote to determine which integration repo we need
to handle.

Original-Bug: 326307398

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1059533
Original-Revision: 8ed9af5eae30194cc04b314546bdbdf9f385be83
GitOrigin-RevId: b8ea6d94a97fc8cd364e38bb2fede380b8363435
Change-Id: Ic7bab43c7a04f9b5b4a8276af66b288031a91935