[roll] Roll fuchsia [kernel][scheduler] Refactor FindTargetCpu into a functional pattern.
Apply feedback from
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/1056093 to
change the use of `FindTargetCpu` into a more functional programming
style.

The basic outline is as follows:

 When a thread needs to be assigned to a CPU/Scheduler, the Scheduler
method used is FindTargetCpu.  This is a static Scheduler method
called while holding a thread's lock, but not any Scheduler's queue
lock.  Its job is to find the best CPU/Scheduler for a thread based on
a number of factors (hard/soft affinity, most recently selected CPU
for a thread, current scheduler load, migration cost, and so on).

Prior to the GTL breakup, the global thread lock was held for the
duration this operation, which functionally locked _all_ Schedulers,
and prevented concurrent thread placement or scheduling, but meant
that the "active" state of a scheduler being considered for thread
placement could not change during the operation.

With the GTL broken up, this invariant no longer holds.  CPU/Scheduler
X could seem to be the best choice for a thread T in the moment, but
by the time the queue lock for X is held after the selection is made,
X may have become de-activated, and therefore not longer a valid
choice.

While this is extremely rare (Schedulers are de-activated *much* less
frequently then they are selected for thread placement), it can still
happen, and the proper remedy is to simply try again.  There must
always be at least *one* valid choice for a thread or other
system-wide invariants are being violated.

As it turns out, this retry pattern shows up rather frequently in the
code (something like 5-6 times).  So, to make this pattern a bit more
automatic and hopefully a bit less prone to copy-pasta mistakes,
refactor it into a more functional pattern which makes use of a
callback-lambda.  A new static method (`FindActiveSchedulerForThread`)
is introduced which handles the boilerplate of attempting to find a
Scheduler, then locking that scheduler and automatically retrying if
the scheduler has become deactivated.  It also provides a central
location where recording retries (done in this CL) or even OOPS-ing if
the number of local retry counts becomes excessive (not done in this
CL, but easily possible in a future follow-up CL if desired).

One of the side-effects of this refactor is that the static locking
invariants become a bit more of a pain to enforce owing to the fact
that Clang's static analysis loses the static knowledge it needs
because of the use of the functional pattern.  Going into the
operation, we would like to statically assert that a thread is locked
and not a member of any container (WaitQueue or Scheduler), and by the
time we make it into our lambda, this is still true, and now we have a
candidate Scheduler which is also locked and active.

A helper method (`MarkInFindActiveSchedulerForThreadCbk`) which tells
the static analyzer what it wants to hear without any extra runtime
overhead has been introduced, and shows up in code as follows:

```
// Requires that our thread be locked
FindActiveSchedulerForThread(..., [](...) {
  // Tell the analyzer that our thread and target scheduler are
  // locked, and the target scheduler is active.
  MarkInFindActiveSchedulerForThreadCbk(*thread, *target);
  // Placement actions go here.
});
```

`FindActiveSchedulerForThread` guarantees that the requirements are
met, and hopefully is sufficiently readable.  If more runtime
guarantees are required, the `Mark` pattern can be promoted to an
`Assert` pattern at a future point in time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1064307
Original-Revision: 4fe9a0fd62d659bed96aea52f5884f97df14283f
GitOrigin-RevId: 7302f37c920bb93aa57bfc829ac4980aad72d5b6
Change-Id: I3c5cd8dee7b520d0cec13cb57a1b014c8538fbe6