[roll] Roll fuchsia [bazel-sdk] Allow disabling package validation.
I'm soon going to increase the strictness of ABI compatibility checks,
and it'd be good to have an escape hatch in place if a product owner
needs to turn these checks off.

Turns out, assembly already has a flag for downgrading package
validation errors to warnings, so this CL simply plumbs that flag up to
the bazel level.

Theoretically, we could have a more granular check - either checking
just a subset of packages, or disabling just these stricter ABI revision
checks, but ideally this flag won't actually be used. If it turns out we
need more of a scalpel than a hacksaw, we can add one in the future.

Original-Bug: 42082683

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063512
Original-Revision: b6863aa8e6415842ebceb85e3c95ce97efb6c137
GitOrigin-RevId: 593dca74f56101b56916f811944fb428d6a3c6b0
Change-Id: I8ddcd58e42a2fd4ce9d304347d4d0d5f23394906