[roll] Roll fuchsia [display][coordinator] Replace display_mode_t with DisplayTiming.
This replaces the display_mode_t used in DisplayInfo for hardcoded
display timings, with the api-types-cpp DisplayTiming struct type.
This change also unifies the returned value type of Controller::
GetPanelConfig() and renamed it to GetDisplayTimings().

The only behavior change is that it's now valid to call SetDisplayMode()
on displays with a fixed display mode, as long as the display mode
in the FIDL call argument matches the fixed mode. The coordinator
will treat this as a no-op, so there'll be no difference in display
configs received by the display engine drivers.

on core.vim3:
Test: "display-tool squares" on vim3 with Amelin AML028-30MB-A1 (HDMI) works.
Test: "display-test --mode-set 1 1" on vim3 with Amelin AML028-30MB-A1 works.
Test: "display-tool squares" on vim3 with Amelin AML028-30MB-A1 works
      after device unplug and replug.
Test: "display-tool squares" on vim3 with Khadas TS050 (MIPI-DSI) works.
Test: "display-test --mode-set 1 0" on vim3 with Khadas TS050 works.

Test: "display-tool squares" works on core.nelson.

Original-Bug: 42078780
Original-Bug: 341156315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057126
Original-Revision: 855ee0b00ab7cfa4e554adf1541e7d25d6ab345c
GitOrigin-RevId: 91ea652492bfd071e931fc6b3e6df52cb8c87357
Change-Id: I979f597fb2603ae2b681b3fa0c5779ed94b58320