[roll] Roll infra/recipes [git] Only check last paragraph of commit msg footers
Previously we would consider any line in a commit message as a candidate
to contain a footer like `Change-Id`. However, this differs from the
officially documented meaning of a footer in git, which is a footer that
appears in the last paragraph of the commit message.

This was causing roll-comment to mistakenly skip commenting on some CLs
because a footer before the last paragraph matched one of the footers
for which it skips commenting.

For example, http://tqr/852048 was a roll of a fuchsia CL
(https://fxrev.dev/1065552) and so the fuchsia CL should have had a roll
comment added to it, but it didn't because the integration CL contained
a `GitOrigin-RevId` footer, but *before* the last paragraph. The roll
commenter skips commits with this footer because it indicates Copybara
commits that won't have a corresponding Gerrit CL to comment on, but
that's only the case of the footer appears in the last paragraph.

This should cause the roll-commenter to start commenting on these types
of changes.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1065812
Original-Revision: 00865b56165ff7aa065b97663cddda0f412ffc55
GitOrigin-RevId: 1cee178ffec1ee7d0a4b5be08bba4a0209a80260
Change-Id: I3057d8c2e0cd8b1bd953427c92878c4842472889