[roll] Roll fuchsia [reachability] Improve Component Exiting Flow
It was possible for the reachability eventloop to exit, which is most
frequently caused by losing connection to the netstack and the watcher
channels closing. Instead of using the ? operator throughout the `run`
body, replace with a call to std::process::exit so that the logs don't
have a reachability panic. Since std::process::exit doesn't output a
stack trace, the ambiguous source of `exit_with_fidl_error` has been
migrated to `exit_with_error` and takes an anyhow error that has a
string explanation of what exactly caused the eventloop to exit.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048470
Original-Revision: 76c1af6cf6ef6255faa3d213363adcb2ace5ac9b
GitOrigin-RevId: 16b44d8eca85a1a2f475f02f869d5eabfee29616
Change-Id: I6aaee22f0b09008e846f7f84a99f5279a82ec574