[roll] Roll fuchsia [netstack3] Implement SO_ORIGINAL_DST
Support the SO_ORIGINAL_DST socket option for TCP sockets, to allow
callers to retrieve the pre-NAT destination of a connected socket. This
option is valid for both IPv4 and IPv6 sockets, but for IPv6 sockets, it
will only return success if the socket is a dual-stack socket and the
connection's original destination was an IPv4 address.

Note that we inspect the TCP socket connection state and the connection
tracking state under the same lock to avoid a race condition.

Original-Fixed: 338042280
Test: netstack-socket-integration-test
      netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059293
Original-Revision: acecc8d14211460807aef016923c94e697884b21
GitOrigin-RevId: d59d6c3b3b2ae4d85ece93df3ad756263a7c9413
Change-Id: I0d4f13efeca8f340ec58d9f8cb73bd35796dae4f