[roll] Roll fuchsia [sensors] Loosen constraint to fix flakey test.
In retrospect, one of the constraints I put on the statistics in
the playback component integration test was not entirely
reasonable. The test has flaked a couple times on this constraint
and I believe it is safe to remove it. The other constraints should
be sufficient.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065293
Original-Revision: 2df34962c89d4234093717fc97b3b0173dba3494
GitOrigin-RevId: ed64b7729387e2a82d5fb4bc5699c8bdec0a4b0b
Change-Id: Ic18bb37bc199fc2c69ea5ea1a8ed88283d74e317