[roll] Roll fuchsia [tee] Refactor the way API implementations are provided
This refactors the TEE Internal Core API implementations and adds some
extra safety to the bindings pattern. There are now 3 layers in the
implementation of TEE bindings:

- tee_internal::binding contains the bindgen generate Rust definitions
  of the TEE internal APIs
- api_impl::binding_stubss contains hand written mechanical forwarding
  stubs that implement the bindgen extern "C" functions by calling a
  (non-extern) Rust function in the api_impl crate.
- api_impl has our implementations of the entry points as regular
  Rust functions within the runtime.

The stubs layer exists so that we can get Rust type checking on our
implementations as the generated extern "C" functions do no type
checking on parameters and return values. This also adds new tests to
check that we are exposing the entry points that we expect to be.

This is similar to https://fxrev.dev/1048382 except that this changes
does not include the memory allocation implementations which need some
more work.

Original-Bug: b/332956721
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065242
Original-Revision: 27d9cd246f3237b90069a7ac8fa8e30082f56685
GitOrigin-RevId: f348feb7a21777d7c07c49727d144def662ad436
Change-Id: I51fdaead9408bd8299386ce246d8745c0c8130ce