[roll] Roll fuchsia [Netstack3] Use ResolvedRoute::local_delivery_device
resolve_route_to_destination() returns local_delivery_device as part
of ResolvedRoute. That field is intended to specify a device for
loopback delivery, but it was ignored during packet delivery. Now
it's passed to the loopback device as expected.
This required some fixes in resolve_route_to_destination - it was
using SRC IP to resolve the delivery device, which is incorrect.
Also renamed parameters in resolve_route_to_destination to make it
less error-prone.

Original-Bug: 42084713
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063879
Original-Revision: 64016549839b8fb6612ed2be44e54c363af86488
GitOrigin-RevId: b6eac5b99196695e89683b1e2d63c9d5f4bc7b7b
Change-Id: Iadd9ab8a2f8264ab3a02a8cd88301d56600f6b01