[roll] Roll fuchsia [fidl][driver] Mark driver.development.Manager server=platform
This is needed to indicate that adding a new strict method doesn't break
ABI compatibility. It says that we only support ABI compatibility in
cases where the servers of this protocol are in platform components.
Clients may be in platform or external components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068293
Original-Revision: 6c83a6613b09e8c57dbc3dd025628e38b94b4455
GitOrigin-RevId: d132c7219fc3d314ad782653d41ff941d2ff879f
Change-Id: I11b962a3ecd9ebe92f26aa6875a77b74513795e0