[roll] Roll fuchsia [audio][drivers] Optional strings must be non-empty if present
For optional strings/arrays to allow BOTH 'empty' and 'missing'
seems the same type of foot-gun as having +0 and -0 be distinct
values. If optional strings don't contain useful content, why
not have them be a predictable value.

This CL clarifies this stance, in the driver FIDL APIs and in
the `audio_driver_tests` suite. Note: this is already being
enforced in the validation section of `audio_device_registry`.

Test: audio_driver_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068516
Original-Revision: 963b9020f139b72662a489c31fbd179c1b372485
GitOrigin-RevId: ca9a9c213d2f3228487d2d264e7def5ee398a29c
Change-Id: Ia225cbe9cfd689200206fce74f9a2ef9328ea368