[roll] Roll fuchsia [sandbox] Open is externalized as a token
Add Open as a dedicated type in fuchsia.component.sandbox/Capability.
The type is represented by a token. External clients can't do anything
with the token besides duplicate it, pass it, or put it into a
Dictionary. We aren't adding any external APIs to actually use the Open;
to be usable, it has to be converted back to its internal
representation.

The purpose of this change is to ensure that, when a Dictionary that
gets externalized contains an Open, it is represented correctly so that
APIs that consume the capability (for example, namespace creation) yield
the correct result. Right now, when Open is externalized, it is turned
into a Connector. This is a lossy transformation because metadata about
the underlying DirEntry type is lost. Also, if the externalized
capability is internalized again, you get a Connector rather than the
original Open. More details can be found in the linked bug.

This may seem like a small step backward in terms of SDK readiness
because it re-adds the external Open type was removed in a previous
change (https://fxrev.dev/1029295). However, the type we are adding back
is a more restricted version, and this is the only short-term approach I
know of that fixes the bug.

Original-Bug: 345815599

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063695
Original-Revision: 08f4799b871913c7a764cac27cd2fa52384be534
GitOrigin-RevId: 325722a755e1a120873c931c7f5dce3984cc6fc5
Change-Id: I2f7461793e3b11d1f1bccd36da6fdfc2ef823e90