[roll] Roll fuchsia [build] Use new "custom-runtime" toolchain tag
This marks a variant like ubsan, that doesn't have
"instrumentation-runtime" because everything in a special
environment should wire up a custom runtime (in the case of
ubsan, via //src/lib/ubsan-custom) if it's feasible.  Things
where it's not feasible to usefully wire up any runtime can
exclude the variant on the basis of "custom-runtime" as well as
excluding "instrumentation-runtime".

This is now used for the things that are built in the user.basic
environment but cannot wire up ubsan-custom.

Original-Bug: 348678677
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065936
Original-Revision: 6630a5246a77a752967e6706088bd929eb4278cc
GitOrigin-RevId: e0f9c21577dcfd9b5b220d02bbd7af3517c12320
Change-Id: I44d23f7b16a7f03a66c137da03cb733c02ef47a4