[roll] Roll fuchsia [test][core] Correct cpu stats test
This corrects a test that is attempting to validate that the number of
records of the ZX_INFO_CPU_STATS can be queried without needing to
provide a buffer. However, the test was doing this by providing the
nullptr for the buffer with a non-zero length. This happened to work
due to the passed in length being less than the size of a single record
and so the kernel not actually attempting to perform any copies to
realize that the buffer was invalid.

Ideally the kernel would validate every passed in buffer range, but
this cannot be efficiently done, and the test was therefore relying on
what is essentially undefined behavior. This commit changes to a
correct buffer size of 0, which will allow the call to reliably work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069572
Original-Revision: 6eb1b84594217c1e3b3a6dfeefb242f5098f98a9
GitOrigin-RevId: d7dff3e64f634e950fd3db36218f46e2f61b1d4f
Change-Id: Ib8830231098d90dfd99169837499646d590803f5