[roll] Roll fuchsia [netstack3] Fix shutdown on connecting sock race
When we move to the closed state as part of shutdown we must remove the
socket from the demux.

The bug was caught when the loom test herein was written to chase a
_different_ race.

Spread some `#[track_caller]` along the way to help debuggig failures
and also teach the ip_test macro about test_matrix.

Original-Bug: 349404597
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071894
Original-Revision: c24f3934bbf6649ff7868740917fb189a5f9e067
GitOrigin-RevId: acb1c0d0cb54fdb90fc2fc5ced602df06545fdaa
Change-Id: I6cdd5ed6fd9771f533ea58fc44d23b78c78221fd