[roll] Roll fuchsia [kernel][phys] Introduce memalloc::Pool::RestrictTotalRam()
This method will be the backbone of the (re)implementation of the
kernel.memory-limit-mb boot option. It removes all accounting for RAM
(free or allocated) past the provided capacity. It is a destructive
routine; removed memory can not be recovered or reallocated.

It is always the tail of RAM that is removed. If allocations are
removed, then that memory is simply forgotten about by the pool. This
applies to bookkeeping to an extent as well: the tracking of bookkeeping
allocation may be removed, but the pool might still have access to
bookkeeping in forgotten ranges, possibly continuing to draw from it
rather than allocate new tracked bookkeeping ranges. This is okay, as it
is assumed that memory that the pool does not track will not be used by
any program.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070845
Original-Revision: 88453ea53325d3658e7a89c7877ca1bb7f9a3d30
GitOrigin-RevId: f19d425147387c39fc765735321fff1b37b51574
Change-Id: I9a55842b7afaa03e46fe5c028f02c300e9120fdd