[roll] Roll fuchsia [zircon] Clarify state of FIDL library `zx`
Attempt to make it clearer that the library is not experimental and is
in a stable but interim state.

In addition, clarify the difference between the two `zx` libraries,
including adding a README for the `zx/` subdirectory.

Also update and fix links in the existing README.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051240
Original-Revision: d62f56ba486d942a224232c55445cccbf17f8e87
GitOrigin-RevId: ece7f70bafd61ed5deeb95fbda23faa570aa2772
Change-Id: If36aa8337a1059519487b28c17351b7d8bb612c3