[roll] Roll fuchsia [starnix][kernel] Clarify get_waitable_child result
Previously, this function returned a `Result<Optional<...`, which made
it unclear what the difference was between an error result and an
`Ok(None)` result. This CL changes the function to return a custom enum
that makes the `ShouldWait` and the `NoneFound` cases clear.

This clarity is especially useful for the one caller, which previously
intercepted the `Err(...)` result and sometimes did not generate an
error (i.e., in the case where it had a tracee to wait for).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1072852
Original-Revision: ced2a359d19a79b9eb75c9b282dae79c74262835
GitOrigin-RevId: 3e927219682e1363a9b48ba4f1f9cf7ae1ffc4f2
Change-Id: I3da5f9ff80dbfc73706c79472306277c2b58507b