[roll] Roll fuchsia [aml-uart] Request interrupt based on driver config
The driver registers the interrupt it requests as wake source without
considering the driver config, with this change, the driver will only
register the interrupt as wake source when suspend is enabled in driver
config.

Test: Verified bluetooth functionality on Smart Display and OOBE worked
well, verified suspend and resume on VIM3 and it worked as expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073652
Original-Revision: 4756bcdf06685616070654e6773834a45b599e9f
GitOrigin-RevId: 6f5af7ac2f89dee0cb054934510f9bacaad66be3
Change-Id: Ib2a8c604578dcef7ba48d359a7e97a3dfdf1ba04