[roll] Roll fuchsia [kernel][phys] s/"extended"/"allocated" in libmemalloc
We initially dubbed the allocated range types as "extended" owing to
their existence as those that live in the extension of the
zbi_mem_type_t value space. This is less clear of a name though; better
to just call them "allocated", especially as we begin to start using
this concept in the kernel proper.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071902
Original-Revision: 57bdf001580ef6d4eeae4f79bd371377cd9123ad
GitOrigin-RevId: 1b7458e1331acae967cb4301512fcab80c39be90
Change-Id: Ic2e5d2584c41ffec3fc3b6efd44c9356dfceced2