[roll] Roll fuchsia [versioning] Replace bump-related asserts with messages
https://fxrev.dev/1062737 added asserts to enforce the conditions that
must be true in order for the simplified implementation to be correct.
Although these are true during the build, they are not true if the GN
rules are run after the build; specifically, after the bump build adds
the new level to version_history.json.

fint-based builds, including platform-version-roller, do this when they
call gnCheckGenerated().

Convert the assert to a warning message that indicates something is
wrong but can be ignored on the bot in this case. The need for these
will be eliminated in upcoming CLs related to the NEXT API level.

Original-Bug: 305961460
Test: `fx set ... --args=bump_api_level=true && fx build && fx gn check out/default`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075392
Original-Revision: a434894d99a07c468eca31dbb97e88b0d4162d53
GitOrigin-RevId: fc5afb16ee0c01dc62518bfd11fa4741b8e634c4
Change-Id: I4b5ec11c30268eae3a73fc74883491064557e4de