[roll] Roll fuchsia [kernel][vm] Correct PmmArena assert
Corrects an assertion in validating the page array allocation of an
arena that was checking if the found range was <= the needed size,
instead of >=. That is to say, it should be considered an error for
the allocated range to not be large enough to hold the desired size.

In practice this is not a bug as boot_reserve_range_search either
returns an error or the exactly passed in requested size.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074892
Original-Revision: e9ad2057b694da63a51b7582f884a891e69601c7
GitOrigin-RevId: 390f08454f0103d314b205747432f3387a7410fa
Change-Id: Id8dd7ca38ab1cf40ab8da912471c762496457899