[roll] Roll infra/recipes [bloaty] Renaming prohibited property "repository" to "remote"
"repository" is now a prohibited property name. This patch renames the
existing property "repository" in the bloaty recipe to "remote". This property is not set by an existing builder that uses the recipe bloaty, so a soft transition is unnecessary.

Original-Bug: 349462629
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1075892
Original-Revision: 16b595a5c0ebf8caa7ef571348db2a96b6b10e79
GitOrigin-RevId: 9ffa9e530243d4e4148d6643f608104d72f07b9d
Change-Id: Ie7e9bf25019d0f58cb5dd3d516573646b46424cb