[roll] Roll fuchsia [netstack3] Skip flaky syscall test
RecvBufLimits is very flaky on netstack3 once we enable multithreading,
and that's because there doesn't exist an appropriate signal to
synchronize this on, so we can't really properly observe dropped packets
from the POSIX API.

This test effectively races loopback delivery with reading from the
receive queue, and if reading from the receive queue wins even one of
the races the test fails.

This was observed as a flaky "Success" when enabling multithreding on
Netstack3.

Original-Fixed: 349938516
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1075222
Original-Revision: 9d9dcf5b041f11f2383b41adbb0efbb6b7904b2b
GitOrigin-RevId: 905478109110765fa4cb577d21cdf7fe4b6b6652
Change-Id: I487425b423446431cc129eeacac769a6d21af3cd