[roll] Roll fuchsia [amlogic-display] Use DFv2 logging implementation in amlogic-display-test.
This change replaces the DFv1 logging backend with DFv2 logging backend
in amlogic-display-test.

For non-driver tests that log to the global logger, a `fdf_testing::
ScopedGlobalLogger` is created  must be created and registered as
the global register to enable DFv2 logging.

Test: amlogic-display-test
Original-Bug: 349404957
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1071972
Original-Revision: 115ac278acca0900cf04bc4214926b60adbd9fb3
GitOrigin-RevId: 1d691ecdf855905c05d6a3ed8b37cea3ada47bf2
Change-Id: I1a6a181f26060fcbc1b2814b26306cff56950e4a