[roll] Roll fuchsia [virtio-vsock] Use ZX_SOCKET_WRITE_THRESHOLD on Rx
We might be observing a race where the application hasn't read enough
data from the socket by the `1msec` timeout and the connection ends up
in an indefinite stalled state in the current implementation.

Instead of a hardcoded wait, modify a threshold signal for observation
in the zircon socket when needed, so we can update the host with the
available socket space.

Original-Fixed: 300933744
Original-Bug: 345795943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1069787
Original-Revision: 7460c09c73e8373bcee3bea44933e65d839f8c55
GitOrigin-RevId: bb70ffe946c9b6bb4a82eeaef095730ae87dc4f1
Change-Id: I163d96962e08efe8d4d1b10bb6c8e43973cd0fdb