[roll] Roll fuchsia [inspect][benchmarks] Rename fuchsiaperf files so that they're not picked up
Rename the "fuchsiaperf.json" files produced by this test to
"fuchsiaperf.fyi.json" so that they are not picked up by the
performance metrics system.

This is the same as what we did for a similar case of another test:
https://fuchsia-review.googlesource.com/c/fuchsia/+/1066032

These tests produce fuchsiaperf files, but they are not fully hooked
up to the performance metrics system, so the results don't get
uploaded to Chromeperf.

This change should unblock (re-)enabling a check to prevent a
reoccurrence of https://fxbug.dev/343997506.  The check had to be
reverted because we got duplicate results from this test on the
"core.x64-asan-nuc_in_basic_envs" builder, because it got run in both
the NUC and AEMU shards (see https://fxbug.dev/350779634).

Original-Bug: 343997506
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076867
Original-Revision: 732ec3a94beada5a8bc56a0dcee7b56a5d9fa00d
GitOrigin-RevId: 1063737b4e2036d80d1422fe4498d802f3f7fd8f
Change-Id: I00e3627045c7745fe9f022dcd6be967edf99da06