[roll] Roll fuchsia [starnix][memory] Attribute memory for thread groups whose leader is killed
Previously the code relied on `thread_group.leader` being a running
task. But that may not be the case in certain edge cases. The new logic
iterates through all tasks in the thread_group to find the `mm`. In the
vast majority of the cases the iteration should finish in the first loop
since that's the leader.

Original-Fixed: b/349117051
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1074475
Original-Revision: 174428226a532bd48a58beaca338a6970c185e5a
GitOrigin-RevId: 25e6fdcaccc0a02aa9b4c3eb0932563ba1198ee0
Change-Id: Ie6f75d70992590aee86cc1371a5ec920354e5e71