[roll] Roll fuchsia [starnix][aio] Implement iovec for async I/O
We now implement the READV and WRITEV operations. Also, improve the
error checking for buffer length and offset to pass more gVisor tests.

Skip AIOTest.BasicWrite because this test assumes that the aio context
is a pointer. Our aio context is just a number, which causes the test to
crash. The aio context being a pointer is related to an io ring feature
that we do not implement yet.

Test: BadIOVecs, BadOffset
Original-Bug: b/297456438
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076862
Original-Revision: 6e655c3bf0cb1300a7300edbf34ad1a3681ca326
GitOrigin-RevId: cf69d1c7c4da4e0ca4aa52a5c6ae13f504a36493
Change-Id: I662c06d57246ac31ba3a02c4c5f961cf95c874a6