[roll] Roll infra/recipes [gerrit_auto_submit] Respect comments from Gerrit
When a human uploads a new patchset of a CL or adds a reviewer/CC,
Gerrit adds a single comment like "Uploaded patch set 6" with a tag like
"autogenerated:gerrit:newPatchSet". As far as I can tell, all tags
starting with "autogenerated:gerrit" are made directly and immediately
in response to human actions on the CL, so they should be considered
when determining the last time the CL was touched by a human.

This should fix the case where auto-submit was incorrectly skipping
"stale" CLs that were approved several days ago but were not submittable
until the owner resolved a merge conflict by uploading a new patchset.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1077873
Original-Revision: c86129756a3efbf20883e72aa33e376c1536efbd
GitOrigin-RevId: dc17f0bba551067a869dd2b65646fcb7119d77cf
Change-Id: Id40e99c52f08085810df48180989250a641b697d