[roll] Roll fuchsia [fx][qemu] Removing build logic from `fx qemu`
This removes the custom logic in `fx qemu` that attempts to only
build the artifacts needed to run the emulator.

As we are making changes to the build system, this logic
continues to be fragile and prone to unexpected results. The better
approach is to use the build system and have developers run
`fx build && fx qemu` to make sure all the relevant artifacts are up
to date before starting the emulator.

This also removes the `--no-build` flag since that is now the only
behavior.

Original-Bug: b/346574369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077852
Original-Revision: 2ae51402f30e6bdf7836e12dd43594884188c1f9
GitOrigin-RevId: e790710328ca76d8b981e5007d6b7e94c13a57c2
Change-Id: I4e3c0cec6fd41c07462fdde8203808895dd4f129