[roll] Roll fuchsia [starnix] Introduce MemoryObject
This is currently abstraction over a vmo used by the vfs to provide
memory to map.

This will allow to describe the content of a file with multiple vmos.
This is required to implement the RingBuffer map type for ebpf

Original-Bug: b/323847465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066352
Original-Revision: e7d4aa9b5456bd4be3a689423d7d35596f9e25a3
GitOrigin-RevId: 68d82b953daa9940611ce2fe1d64cf628446826b
Change-Id: I08d1271c40770c074c41bce491adcb3482bc3ef3