[roll] Roll fuchsia [docs][serdedoc] Adding md_escaped macro in handlebars
This adds a macro to use in the handlebars templates to
escape characters that confuse or break the markdown parsing
of the output.

These characters are things like [] () {}, etc.

This is done in preparation of running doc-checker as part of
the reference doc generation pipeline.

Original-Bug: b/344621741
Original-Bug: b/346408156
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077901
Original-Revision: d7965c391468da1c572aab12378bcae374268319
GitOrigin-RevId: 882fa7adefcf262d1d0e648a4023d7ce038faa3b
Change-Id: I680bb45eaef026e1bc122e7befac6dd69797e6dc