[roll] Roll fuchsia [ffx] Fix dynamic adding of target to daemon
OpenTarget(<addr>) should add the target if it doesn't already
exist in the TargetCollection. But the newly-added target wasn't
getting added with the specified address, so the OpenTarget actually
failed to find the target to ask it to establish the host-pipe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078005
Original-Revision: 477a1fc72d7e3f1f2c8a3283e9d12f6202ed19f3
GitOrigin-RevId: 0e82e34b656d268b7b2220e3d557c3106916d93a
Change-Id: Ibaa18acbf835ccd2977f8894bf7609dda8e40c21