[roll] Roll fuchsia [bt][avrcp] Fix avrcp relay incomplete setup issue.
Sometimes, peers do not respond to our initial AVRCP requests to get the
media attributes/play status and this causes us to hang and not send a
WatchInfoChange response to mediasession.
Implement a work around so that place a time-out on the requests,
so that it doesn't block us from setting up the relay.

Test: fx test bt-a2dp-tests
Original-Bug: b/343196880
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073538
Original-Revision: 02c39294c2a840afd95295c7dbbe623d6240729e
GitOrigin-RevId: e26d7cfa486f2ff586f5e44e0c3ea3012d716ce1
Change-Id: I030c11b520581fa53c930acdf2ec79b1fc63fc63