[roll] Roll fuchsia [ffx]Removing unreachable help ffx_plugin
This removes the ffx_plugin for Help in the ffx frontend.

It is unreachable since argh handles the "help" subcommand.
The jinja template used to chain together all the plugins and subtools
still expects the top level of ffx to be a "plugin suite", even if
the code is unreachable, so some boilerplate is needed until the
jinja template is edited or removed.

This is the last instance of ffx_core::ffx_plugin.

Original-Bug: b/290067014
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078795
Original-Revision: 265d90d956323c0d32f21af19b24d6419fee778d
GitOrigin-RevId: c424a071ec82bc0a34cb910dc5ab825ab0500b9c
Change-Id: I2de00fc35a6fb78695c314e2ffe3304a46749ee1