[roll] Roll fuchsia [devicetree][gpio] Make gpio-init bind to a specific controller
Make composite node specs with a gpio-init parent bind to the specific
GPIO controller that is responsible for executing the init steps that
the composite node depends on. This ensures that the initialization
order is correct on boards with multiple GPIO controllers.

Test: fx test gpioimpl-visitor-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060144
Original-Revision: a79a686f8a36b0ccbbb31d1d13edae6ee0ce14cf
GitOrigin-RevId: dce2b24bd664903a201af7cc31cc1ecbf564dd52
Change-Id: Ia7cfd45c7bd7ca6214bd3640a5eaabd0f64734d2