[roll] Roll fuchsia [starnix][pty] Send signals to the correct process group
We are only supposed to generate signals when transforming the terminal
input queue and those signals are always delivered to the foreground
process group of the controlling session for the terminal.

Previously, we computed signals both for the input and output queues,
and we maintained two session associations, one for each end of the
terminal. None of that complexity appears to be correct. This CL
simplifies these data structures and passes more tests.

Additionally, the terminal was maintaining its own concept of the
foreground process group for a session, which appears to be incorrect.
The notion of a foreground process group is scoped to the session
itself, not to a given terminal.

Test: JobControlTest.ControllingProcessPersistsAfterChildExists
Original-Fixed: b/42159560
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078792
Original-Revision: 058d17dd51f221ad5ac8b4ae926f6309d33d1b38
GitOrigin-RevId: 889a3a7a23bc2fc078aa6c93e61743db3e4f4cff
Change-Id: I4a4510b28027433092df2cca5f89b3c6687d8420