[roll] Roll fuchsia [debugger] Improve help text when failing to infer filenames
The previous text was not particularly helpful:

```
[zxdb] ...
There is no current file name to use, you'll have to specify a file.
```

Now we give better instructions for what to do:

```
[zxdb] ...
There is no file name to use in the current context. Either specify a
`pause`) and the file that you want to stop in is the active stack frame with
`frame` to use line numbers with a contextual file name.

See `help break` for details.
```

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079632
Original-Revision: 5b040aeec9eed0129596b3adddf27dcbe3e5cf05
GitOrigin-RevId: ed612842768c45f7c371b553cf29fdb0c9e3b46e
Change-Id: If9f40accb9c87c0617ac71cd65ac71bbe2069505