[roll] Roll fuchsia [starnix][gvisor] Skip AIOTest.BadWrite
This test is flaky because we generate the error that the test is
looking for asynchronously, which means the test passes when we're fast
enough to generate the error before the test checks for it.

The real fix is to rework our async I/O code to do more work
synchronously, but, for now, we just skip the test to avoid the flake.

Original-Fixed: b/352233059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080132
Original-Revision: 7815cf753477e0ecaa04b1093569d6bfeeb28034
GitOrigin-RevId: 77a98b6464eb9c633368b8ee3c133a3e7d211cee
Change-Id: I828489d5aae8a91f8fa16f1423e04296dd1d10ec