[roll] Roll fuchsia Reland "[dtr] realm factory use test.Capability"
This is a reland of commit 2bcb4080d7570578f27edef5513f1f6a97f6483e

There used to be an entry for driver_urls in the driver test realm's
RealmArgs table at entry 14. This was removed as it was not part of the
stable sdk. Unfortunately we have CTF tests that are frozen that ended
up using that field. This causes the decoding on the realm factory side
to break. Disabling these tests until those branches are updated to
remove the driver_url field so that the tests can be reenabled.

Original change's description:
> [dtr] realm factory use test.Capability
>
> In the driver test realm's start args, we can re-use the
> fuchsia.component.test.Capability instead of creating our own type. This
> will allow us to offer and expose other capabilities instead of being
> tied to the offer=protocol and expose=service. This also changes how
> these are routed in the driver test realm, to directly forward these
> instead of manually serving up exposes. Keeps the old way for a soft
> migration, but marks them as deprecated. These new fields are added into
> the parner sdk.
>
> Tests: basic-driver-test
> Original-Bug: 42073139
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061660

Original-Bug: 42073139
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079512
Original-Revision: 2e1aaac4297c15774d06e54d31e08ce8a5f6cd6f
GitOrigin-RevId: 559defc9acbbfdc871e95e6d17b524af56e167c5
Change-Id: I35485d022104cb02b2da45c6fc46dd3d53341aa4