[roll] Roll fuchsia [debugger] Fix name resolution after dot operator
In specific situations where there is a type that conflicts with a
member name, the expression parser will always take the type before
looking for variable identifiers with that name.

In the context of evaluating the right hand side of a "." or "->"
operator in an expression, types should be ignored. If a type does
happen to conflict with the token that would otherwise identify a
particular member on an object, we should treat the name of the type
like an identifier and then let the member accessor logic determine if
there is actually an identifier there by that name.

Original-Fixed: 352125964
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079633
Original-Revision: 97849a0aa29c56f6ad0c55e618a8ff2d6e6f7b75
GitOrigin-RevId: cff1dc2d8fbf7c375266e7a58740fc75a3accdd4
Change-Id: I60bb122ec79df897ce24f7968198cccd92a391ba