[roll] Roll fuchsia [netstack3] Simplify DeviceSendFrame error
In preparation to better handling of dequeueing frames, simplify the
errors that can be observed when sending frames to the device.

First, reframe the generic "NotReady" error (which is never yielded by
bindings) to "NoBuffers" which is closer to errors that we might
actually observe in practice -- though we're not yielding that just yet
because the tx queue handler in bindings is not going to behave nicely
if it observes that error.

If any errors are observed during dequeueing, we just drop the packet
and increment a counter. Given this packet has already been delivered to
packet sockets we should not be requeueing it again.

Original-Bug: 42057204
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080212
Original-Revision: fc0e7215465508e8dc3ba3fa2c6afe145132fa23
GitOrigin-RevId: e2186f0f7ef82fc2a77a118b521d985d264ad33b
Change-Id: I37b12edd0ec579f7f4a2889fe274d2e770219fdd