[roll] Roll fuchsia [versioning] Add support for NEXT to version_history
Add support for API level "NEXT" to version_history scripts and code so
that they are ready for the switch to "NEXT" as the only in-development
API level.

get_platform_version.py now requires that `api_levels` have a status of
"unsupported", "sunset", "supported", or "in-development", and that
`special_api_levels` have a status of "in-development". This was already
the case in practice but was not enforced.

Original-Bug: 326277078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073565
Original-Revision: 333d0ff10bc4a89002f5eae2f95620e2f46803b5
GitOrigin-RevId: 87f7b9ffed157cf5b8e974ce45f64f876a9cc45c
Change-Id: I5b3cbfd890c8031071d22488fd798b1fc991b025