[roll] Roll fuchsia [vfs][rust] Return EntryInfo for Node and DirectoryEntry
Introduce a new trait `GetEntryInfo` that all implementations of
`DirectoryEntry` (used in pseudo filesystems) and `Node` must
implement. This trait only has one method that returns `EntryInfo`.

Prior to this change, there is a method to get `EntryInfo` but it was
a part of only `DirectoryEntry`. Being able to get the entry's
information, regardless of if it is used in a pseudo filesystem, is
still useful. For example, in Open3, we want to identify the directory
entry type to validate protocols when there are more than one specified.

Original-Bug: 348698584
Test: fx test //src/storage/conformance
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1079792
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: cf65695c3ade53ebb798c9387f5f473aa67def1b
GitOrigin-RevId: 4edaf8bf2c7972b78a471f47af2e4b5b89cfc22c
Change-Id: I8160ec95846c39a429a9474b51f369597893e0d5