[roll] Roll fuchsia [intel-i915] Pass an Inspector on Controller creation.
This change adds an Inspector argument to the Controller constructor and
its factory function. This makes it possible for drivers to pass an
Inspect to a hardware component class, which is required for DFv2
drivers to set up the component inspector.

Test: intel-i915-test
Test: On core.x64 on NUC11, "ffx target snapshot" output contains
      intel-i915 root node.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076537
Original-Revision: a84c88313fe10e79566340203c0545f7e3993fce
GitOrigin-RevId: be9321908e0190ad471cb7a930958449c790b740
Change-Id: I8f41f78006051c59e4247d2f63e779f3c8b8e8b8