[roll] Roll fuchsia [netstack3] Move Segment into base
...along with all transitive dependencies.

We want to be able to reuse this struct as part of conntrack's TCP state
tracking. It couldn't be used directly, since the netstack3_tcp crate
has a dependency on the filtering crate, so we'd end up with a
dependency cycle.

This also required moving a few impl blocks to avoid running afoul of
the orphan rule.

I tried to retain history as much as possible, but a few files were only
partially copied, so their history is sadly lost.

Original-Bug: 328064736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080553
Original-Revision: 37b968ad15e66915de23d08f7cd29864098b7ace
GitOrigin-RevId: da41f8d6ed6b41fbaf98e6d239260c35cbccf5bc
Change-Id: Ib9bce205b5e4c99d9c393fc5df35cf96c75ad462