[roll] Roll fuchsia [bazel-sdk] add build time check for restricted symbols
Adds the ability to provide a set of restricted symbols to be checked at
build time. If the library uses any of the restricted symbols then we
will fail at build time instead of runtime.

Note: We are not actually enabling this check in this CL until we
understand why symbols are being pulled in during RBE builds. Once the
linked bug is better understood we will enable this feature so we don't
regress in the future.

BUG: b/352586714
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081978
Original-Revision: cb31e39131767e352ee9091f5c31ecb44640102f
GitOrigin-RevId: 47d7be31bdd5cdbf2d57196a1b4bc74f4ae7d79c
Change-Id: I645d52f653409c7dc6a960b8f4c3dd18023e4e24