[roll] Roll fuchsia Reland "[power][broker] Create internal lease element"
This is a reland of commit 53a4ea3a2583d15b686b2ef63173ea1efb81f636

The original commit broke the power e2e tests on the VIM3. When the
timer received the interrupt, it would grab a wake lease to keep
the system out of the suspended state and block the dispatcher until
the lease was satisfied. However, this change now does not consider
a lease satisfied until the consumer element acknowledges the level
update request, which it cannot do while the dispatcher is blocked.

In the near future, power APIs will be introduced to avoid the need
for consumer elements, thus eliminating this problem. The workaround
employed here is to not wait for the lease to reach satisfaction. In
the case where execution state is active, grabbing the lease will
activate the execution state claim and keep it active. In the case
where it is inactive, there is a possibility of this racing if the
hrtimer completes (thus ack'ing the interrupt) and the system tries
to suspend in the gap between the interrupt clearing and SAG running
its power elements once the interrupt handler returns. Given the
limited impact of this behavior and the line-of-sight to a better
fix, this cost is absorbed for now.

Original change's description:
> [power][broker] Create internal lease element
>
> Adds an intermediate, synthetic lease element that has an assertive
> dependency to the element being leased. This changes the behavior of
> leasing elements in the power broker to return 'Satisfied' only when the
> element being leased has reached the desired level. Prior to this
> change, the broker would return 'Satisfied' when the dependencies of the
> leased element have reached their required levels.
>
> Original-Bug: 345829869
> Test: fx test src/power/broker
> Test: fx test fake-system-activity-governor-test
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1066613

Original-Bug: 345829869
Cq-Include-Trybots: luci.turquoise.global.try:begonia_eng.vim3-release-power
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081854
Original-Revision: 7b351cbb1625738c1c6d416fc1bf8db3410b81c4
GitOrigin-RevId: ce6e4a8122a736e0f17ff36ef05799d2678902e2
Change-Id: I71e57ea631f87c5adc1003c43d0deba6f91ca8cf