[roll] Roll fuchsia [ffx] Reland dynamic adding of target to daemon
(This was reverted unnecessarily -- the culprit was another CL.)

OpenTarget(<addr>) should add the target if it doesn't already
exist in the TargetCollection. But the newly-added target wasn't
getting added with the specified address, so the OpenTarget actually
failed to find the target to ask it to establish the host-pipe.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082638
Original-Revision: d6d506bc6d625c63ba9bffbd2e951eb8c02b0342
GitOrigin-RevId: 301c478818b5df06a5a93a65ef6b4213e1381914
Change-Id: I1d9b70c603907610bf62257919490c23afe08115