[roll] Roll fuchsia [versioning] Update API level bumping to use NEXT
In preparation for switching to "NEXT" as the in-development API level,
update the bumping and FIDL golden generation parts of the build to
always use "NEXT" as the in-development API level.

The old numbered in-development level model is still supported, and
golden files for `NEXT` are not yet generated. However, `bump_api_level`
now adds a `"supported"` API level and is thus no longer compatible with
the old model.

`fidl_api_compatibility_testing.md` now uses `NEXT` in the example and
will no longer be updated when bumping.

Original-Bug: 326277078
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073566
Original-Revision: 8c2808aa6be8b98e5676acbc25f4f05a3f8e0b3e
GitOrigin-RevId: 9840d4b7d2b7d3f50cfa08c086e8217a722e2f13
Change-Id: Ie25170516f02ce2cfb68b2399e330d5e6d5ced15