[roll] Roll fuchsia [netstack3] Allow Deferred Device Socket Removal
During deferred device removal, there is a point in time where
WeakDeviceIDs cannot be upgraded, but the device's state has not been
dropped. During this window, closing a device socket on the device would
lead to a panic, because the device state still holds a strong reference
to the device socket.

To fix this, add support to the DeviceSocketApi for deferred socket
removal.

Original-Fixed: 353484978

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082054
Original-Revision: 9b25c4f9e05af539c828302cb08f8fda644752b8
GitOrigin-RevId: 2a83bc2b81557cc8a4199fb935d7d01f0507a2e4
Change-Id: I879486423588e88394b1e878fc06c30c5483ed89