[roll] Roll fuchsia [starnix][selinux] Fix the composite policy framework initial SIDs
The set of initial SIDs is defined by the Reference Policy, and
the order of declarations is part of the ABI.

An "initial_sids" file containing the Reference Policy is added,
which must be passed as an explicit argument, separate from the
rest of the policy fragments. The policy merge process validates
that only the "initial_sids" file attempts to define new SIDs.

Original-Bug: b/343891990
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080952
Original-Revision: 16eb5ac1b8a7d443611a7e17a222e6df311d0b3a
GitOrigin-RevId: b200b88682185687d20291f9960be5f65e34f480
Change-Id: Id99af817b056895372b3faf3a51db29eb60bc0e2