[roll] Roll fuchsia [assembly][bt] Correctly add bt-transport-uart driver to assembly
As is, the driver is only included on builds that require "Standard"
bluetooth (e.g. all Bluetooth components). It is valid for the driver to
be included on "Coreless" builds as well, so move up the platform bundle
inclusion before the early return in the config.

Test: Manual
Original-Bug: 354005030

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084662
Original-Revision: 760df92ad84151cb6879831af3da643047353f47
GitOrigin-RevId: a2ae31d48cf58d90e74eceb12cfd72f649232e57
Change-Id: If194292f1c78644527c6a68d1ec5b3084eec12c5