[roll] Roll infra/recipes [checkout] Emit cached_revision output prop earlier
This ensures that the `cached_revision` property gets emitted even if
the checkout fails. This is important for the correctness of the
orchestrator build's determination of whether the subbuild had a cache
hit or not, in the case where the checkout fails.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1085352
Original-Revision: 27f39c13af1109d5bebf8afd3cb856905720745c
GitOrigin-RevId: 2f2ba0892c1c8865afce3d8b7abbed840b8e7e05
Change-Id: Ic8fd038764a60c9c011433ce46179b9b8711eb33