[roll] Roll fuchsia [kernel][sched] Have blocked threads temporarily contribute firm demand
This change introduces an optimization in which threads that become
blocked will contribute firm demand for the remainder of the activation
period in which that took place. This allows threads that quickly
unblock to complete the still-accounted-for firm work they set out to do
in that period and not be penalized, minimizing disruption.

Original-Bug: 328641440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1029434
Original-Revision: 5e0ed7c4b6116ef8996f1d0b91c6cb846e08cb51
GitOrigin-RevId: d4ecfcfa1609a63a404f0eaae7532857f058dd8b
Change-Id: I46a3f8188f54252eca3b843ed3b62fc6bc4438ad