[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Fix GPU 'printf' on strings with padding
Summary:
We get the `strlen` to know how much memory to allocate here, but it
wasn't taking into account if the padding was larger than the string
itself. This patch sets it to an empty string so we always add the
minimum size. This implementation is slightly wasteful with memory, but
I am not concerned with a few extra bytes here and there for some memory
that gets immediately free'd.
GitOrigin-RevId: 247c06653f8993b34a24c4073e9d08ba52adadbd
Original-Revision: d9f700708e866628d74a01d9716ff7eed7919f8f
Roller-URL: https://ci.chromium.org/b/8741821607133538369
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086155
Original-Revision: 32003d219b120bb85160651a59025be9e026c946
Change-Id: I81a3328c7441bd230d3755bfa083d4942719af4e