[roll] Roll fuchsia [assert] Conditionally use [[likely]] in ZX_ASSERTs
Only use the [[likely]] attribute in the various ZX_ASSERT
implementations if the compiler understands it.  The attribute was
formally introduced in C++20, and while many compilers will tolerate
its presence, not all will.

So, use the preprocessor to test to see if the current compiler
supports the attribute, and only use the attribute if the compiler
understands it.

Original-Fixed: 354715786
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086916
Original-Revision: 9df2333aef1349a2c5b349d79cd5ef510b6b64f9
GitOrigin-RevId: 4dbce6fe624da7b64e656faeee3bc36bc523b72b
Change-Id: I8b7b508a498bc7f0bef7577d8b80be7df2f8d04d