[roll] Roll fuchsia [tee] Wire up memory TEE calls
This implements the TEE basic memory management calls. The TEE memory
management APIs can be configured through hints to zero the contents of
new allocations and to keep allocations private from other TAs. We
always satisfy the second requirement and to minimize bookkeeping it is
simpler to always zero contents as well (OP-TEE does the same). To do
this we provide allocator configuration values. We have to provide two
different options depending on if the build is configured to use Scudo
(the default) or ASan's allocator.

Original-Bug: b/332956721

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068236
Original-Revision: dc3951131144959f4c60c3fa94041d9ea106ff4f
GitOrigin-RevId: 9d164d567d990a2b41e58444055e48403877f0be
Change-Id: I22f3161330cd14ed82b503beca5a45cb6633bea1