[roll] Roll fuchsia [kernel] Support boot timers in the kernel
Adds support for timers operating on the boot timeline to the kernel.
This required modifying the TimerQueue to contain two separate timer
lists: one for monotonic timers and one for boot timers.

The reference timeline a particular timer is operating on is declared on
construction of the timer object, and is encoded as an enum member
variable. This may be overhauled in future patches once stronger types
for time in the kernel have been developed.

This is a necessary prerequisite to pausing the monotonic clock, as the
resume timer is our first boot timer. However, because we do not
currently support generic boot timers (e.g., zx_timer_create does not
allow callers to specify the boot timeline), the code asserts that only
the resume timer is ever set.

Original-Bug: 341785588, 328306129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1080638
Original-Revision: 3239f7152b8762560694f27c4e0760a9a8d513fb
GitOrigin-RevId: 5a9c4a9dff48b2850c22e285277c4beb67564145
Change-Id: Iae93ed68c88574cd151bb9cc148482935db3e1db