[roll] Roll fuchsia [tefmocheck] Add check for unable to resolve address.
This message can occur if we fail to connect to the target which can
happen before the test runs and thus would be unrelated to the test.
Adding a tefmo check will help us track this as well as avoid filing
bugs for tests that fail because of this.

Original-Bug: 354707902
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087272
Original-Revision: f001eaf091b739cc57fe775cfa9fdc016df57f66
GitOrigin-RevId: 99978b797279702d6cb12761d749316e004d7838
Change-Id: I8f01ccc2c9c437af3f77f73af583459087b99295