[roll] Roll fuchsia [serdedoc] Adding disconnected types to toc
Due to b/332348955 enums that have different variants are
not connected to the data type graph. As a result, the README.md
for the type is generated, but the _toc.yaml does not include the file.

This situation blocks running doc-checker on the generated reference
docs. As a stopgap, this change adds the disconnected types to the
bottom of the toc for assemblydoc.

Original-Bug: b/332348955
Original-Bug: b/344621741
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087694
Original-Revision: dc24de0881edbefad4652080440b7fb7927cb1bd
GitOrigin-RevId: 08c2ec81b835e82f8b34e9ed7728c76f9a1ad8c7
Change-Id: I4a5014ad07480accda2a55908f6b05786407dd34