[roll] Roll fuchsia [driver_runtime][rust] `Arena::into_raw` isn't unsafe
Remove the unsafe designation and warning from `Arena::into_raw`, since
it's not unsafe to leak the arena.

Also adds `fdf_arena_t` to the public api of the crate, since it's
exposed in some functions for going to and from raw pointers. It was
missed because it's just a type alias to `u32`, and so did not trigger a
warning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087877
Original-Revision: e7697f82eddf7c7dfad17a4521ca76787c7eb691
GitOrigin-RevId: 3a3b2ab70be6721483f20f6f490278ce3c0e1dc9
Change-Id: I4110fcc36c1c2fbacefa3afdfeae4e1b96411a82